Refactorización de los componentes. Se ha metido todo en carpetas porque me estaba volviendo loco con todo en una única carpeta
