+++ transportes-fernandez/src/components/layout/Header/Header.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Container, AppBar, Toolbar } from '@mui/material';
-import './Header.css'; // Importar el archivo de estilos CSS
-import Logo from './Logo';
-import NavButton from './NavButton';
-import routes from '../data/routes'; // Importar las rutas
+import './Header.css';
+import Logo from '../../features/Logo/Logo';
+import NavButton from '../NavButton/NavButton';
+import routes from '../../../data/routes';
 
 function Header() {
   return (
